Changed storage class of functions in drivers/net/cpsw.c to static Added tab completion for shell Added function to parse ethernet header

+++ lib/net/print-packet.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-void print_packet(void *buf, size_t len)
+void print_packet(void *buf, uint32_t len)
 {
 	int data, temp, i;
